requirements.txt
aiohttp
celery
eventlet
msgpack-python
pika
PyYAML
redis
requests
